Add dbNSFP predictions to VCF
dbNSFP predictions can now be added to the VCF file. In addition to what
dbNSFP already contains, the pipeline will also add a nominal prediction
value for GERP++ and phyloP20way mammalian to be displayed on the DVD
website. Lastly, the total number of predictions available, the number of
pathogenic predictions, and a final (majority vote) prediction are all
added.