Improve ability to handle slight pathogenicity disagreement
Slight disagreement between ClinVar and HGMD is now handled better. For
example, if one says "Pathogenic" and the other says "Likely pathogenic"
then the variant is automatically called "Likely pathogenic", and a note
is made in the commments to let the user know of this slight
disagreement. The note will also state the source of each pathogenicity.