Fix querying of overlapping regions
bcftools will produce unsorted and duplicated VCF records if you
give it overlapping regions. Therefore, overlapping regions are merged
before being passed to bcftools. This means that the genes2regions
method produces two .bed files: a merged regions file that is used to
query VCF files and a non-merged file to annotate a VCF with gene
symbols. This also means that bedtools is now a dependency.